refactor: use flashsale-home styles
