--- react-app/src/styles/flashsale.module.css
@@ -1,49 +1,0 @@
-.flashsale-menu {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  padding: 8px 0;
-}
-
-.flashsale-menu button {
-  padding: 6px 12px;
-  border: none;
-  background: #660066;
-  color: #fff;
-  border-radius: 9999px;
-  cursor: pointer;
-}
-
-.flashsale-menu button:hover {
-  opacity: 0.9;
-}
-
-.flashsale-page .tour-card {
-  border: 1px solid #e5e7eb;
-  padding: 8px;
-  margin-top: 12px;
-}
-
-.flashsale-page {}
-
-.countdown {}
-
-.tours {}
-
-.departures {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-top: 4px;
-}
-
-.departure {
-  background: #f3f4f6;
-  border-radius: 9999px;
-  padding: 2px 8px;
-  font-size: 12px;
-}
-
-.quantity {}
-
-.selected {}
